move thymeleaf view class to separate file
