+++ modules/thymeleaf/src/main/kotlin/org/springframework/fu/module/thymeleaf/ThymeleafModule.kt
@@ -18,23 +18,10 @@ package org.springframework.fu.module.thymeleaf
 
 import org.springframework.context.support.GenericApplicationContext
 import org.springframework.context.support.registerBean
-import org.springframework.core.io.buffer.DataBufferUtils
 import org.springframework.fu.AbstractModule
 import org.springframework.fu.module.webflux.WebFluxModule
-import org.springframework.http.MediaType
-import org.springframework.web.reactive.result.view.AbstractUrlBasedView
 import org.springframework.web.reactive.result.view.UrlBasedViewResolver
-import org.springframework.web.server.ServerWebExchange
-import org.thymeleaf.TemplateEngine
-import org.thymeleaf.context.Context
-import org.thymeleaf.spring5.SpringTemplateEngine
 import org.thymeleaf.spring5.SpringWebFluxTemplateEngine
-import org.thymeleaf.spring5.context.webflux.SpringWebFluxContext
-import reactor.core.publisher.Flux
-import reactor.core.publisher.Mono
-import java.io.InputStreamReader
-import java.io.OutputStreamWriter
-import java.util.*
 
 /**
  * @author Artem Gavrilov
@@ -48,38 +35,12 @@ class ThymeleafModule(private val prefix: String,
             UrlBasedViewResolver().also {
                 it.setPrefix(prefix)
                 it.setSuffix(suffix)
-                it.setViewClass(MyView::class.java)
+                it.setViewClass(ThymeleafView::class.java)
             }
         }
     }
 }
 
-class MyView : AbstractUrlBasedView() {
-
-    override fun checkResourceExists(locale: Locale): Boolean {
-        return applicationContext?.getResource(url!!)!!.exists()
-    }
-
-    override fun renderInternal(model: MutableMap<String, Any>, mediaType: MediaType?,
-                                exchange: ServerWebExchange): Mono<Void> {
-        val dataBuffer = exchange.response.bufferFactory().allocateBuffer()
-        try {
-            val engine = applicationContext!!.getBean(SpringWebFluxTemplateEngine::class.java)
-            val resourceStreamReader = InputStreamReader(applicationContext?.getResource(url!!)!!.inputStream)
-            val charset = Optional.ofNullable(mediaType?.charset).orElse(defaultCharset)
-            resourceStreamReader.use { reader ->
-                OutputStreamWriter(dataBuffer.asOutputStream(), charset).use { writer ->
-                    engine.process(reader.readText(), SpringWebFluxContext(exchange, Locale.getDefault(), model), writer)
-                    writer.flush()
-                }
-            }
-        } catch (ex: Exception) {
-            DataBufferUtils.release(dataBuffer)
-            return Mono.error(ex)
-        }
-        return exchange.response.writeWith(Flux.just(dataBuffer))
-    }
-}
 
 fun WebFluxModule.WebFluxServerModule.thymeleaf(prefix: String = "classpath:/templates/",
                                                 suffix: String = ".html"): ThymeleafModule {
